Version1.0

+++ W801源码V1.0/app/user_pcap01.h
@@ -25,6 +25,6 @@
 
 
 void pcap01_init(void);
-double pcap01_read_res(uint8_t addr);
+float pcap01_read_res(uint8_t addr);
 
 #endif
